Refactor episode to fragment
