+++ app/src/main/AndroidManifest.xml
@@ -28,9 +28,6 @@
                 android:resource="@xml/shortcuts" />
         </activity>
         <activity
-            android:name=".presentation.activity.EpisodeActivity"
-            android:screenOrientation="portrait" />
-        <activity
             android:name=".presentation.activity.CharacterActivity"
             android:screenOrientation="portrait" />
         <activity
